+++ examples/get-file-contents.js
@@ -15,7 +15,7 @@ async function main() {
         .setFileId(FileId.ADDRESS_BOOK)
         .execute(client);
 
-    console.log(resp.contents);
+    console.log(resp);
 }
 
 main();